Revert "Split `arm-resource-operation` lint rule into three focused rules (#4808)"
This reverts commit a55c28f74db70c86f6477cb0204ef7b6de5b0240.


Copilot-Session: 42e3447b-a983-4ca4-a095-640b7560d244
