--- .chronus/changes/arm-resource-operation-rule-split-2026-07-02-22-17-05.md
@@ -1,8 +1,0 @@
----
-changeKind: feature
-packages:
-  - "@azure-tools/typespec-azure-resource-manager"
-  - "@azure-tools/typespec-azure-rulesets"
----
-
-Split `arm-resource-operation` lint rule: add `use-operation-decorator`, `use-api-version`, and `use-interface` as separate rules replacing the original combined rule.
