--- packages/typespec-azure-resource-manager/src/rules/use-operation-decorator.ts
@@ -1,164 +1,0 @@
-import type {
-  DecoratorApplication,
-  DecoratorArgument,
-  Operation,
-  Program,
-} from "@typespec/compiler";
-import { createRule, isTemplateInstance, paramMessage } from "@typespec/compiler";
-import type { HttpVerb } from "@typespec/http";
-import { getOperationVerb } from "@typespec/http";
-import { getSegment } from "@typespec/rest";
-import {
-  isSourceOperationResourceManagerInternal,
-  isTemplatedInterfaceOperation,
-} from "./utils.js";
-
-export const useOperationDecoratorRule = createRule({
-  name: "use-operation-decorator",
-  severity: "warning",
-  description: "Validate ARM Resource operations use the correct decorator for the HTTP verb.",
-  url: "https://azure.github.io/typespec-azure/docs/libraries/azure-resource-manager/rules/use-operation-decorator",
-  messages: {
-    default: paramMessage`Resource ${"verb"} operation must be decorated with ${"decorator"}.`,
-    invalidOperationType: paramMessage`Resource ${"verb"} operation decorator ${"decorator"} must use operationType ${"operationType"}.`,
-  },
-  create(context) {
-    return {
-      operation: (operation: Operation) => {
-        if (
-          !isSourceOperationResourceManagerInternal(operation) &&
-          !isTemplateInstance(operation)
-        ) {
-          const verb = getOperationVerb(context.program, operation);
-          if (verb && isTemplatedInterfaceOperation(operation) === false) {
-            const requiredDecorators = resourceOperationDecorators[verb];
-            if (requiredDecorators?.length > 0) {
-              const decorator = operation.decorators.find(
-                (d) => requiredDecorators.indexOf(d.decorator.name) >= 0,
-              );
-              if (!decorator && !isArmProviderOperation(context.program, operation)) {
-                context.reportDiagnostic({
-                  messageId: "default",
-                  target: operation,
-                  format: {
-                    verb: verb.toUpperCase(),
-                    decorator: requiredDecorators.map((d) => `@${d.substring(1)}`).join(" or "),
-                  },
-                });
-              } else if (decorator) {
-                const allowedOperationTypes = operationTypesByVerb[verb];
-                const operationType = getGenericDecoratorOperationType(decorator);
-
-                if (
-                  allowedOperationTypes !== undefined &&
-                  operationType !== undefined &&
-                  !allowedOperationTypes.includes(operationType)
-                ) {
-                  context.reportDiagnostic({
-                    messageId: "invalidOperationType",
-                    target: operation,
-                    format: {
-                      verb: verb.toUpperCase(),
-                      decorator: `@${decorator.decorator.name.substring(1)}`,
-                      operationType: allowedOperationTypes.map((x) => `"${x}"`).join(" or "),
-                    },
-                  });
-                }
-              }
-            }
-          }
-        }
-      },
-    };
-  },
-});
-
-// These private ARM decorators cover all operation types and serve as valid alternatives
-// to the public per-verb decorators (e.g. in extension resource, legacy, and built-in operations).
-const genericArmResourceDecorators = [
-  "$extensionResourceOperation",
-  "$legacyResourceOperation",
-  "$builtInResourceOperation",
-  "$legacyExtensionResourceOperation",
-] as const;
-
-type GenericArmResourceDecorator = (typeof genericArmResourceDecorators)[number];
-
-// operationType argument position per decorator signature:
-// - extensionResourceOperation(targetResourceType, extensionResourceType, operationType, ...)
-// - legacyResourceOperation(resourceType, operationType, ...)
-// - builtInResourceOperation(parentResourceType, builtInResourceType, operationType, ...)
-// - legacyExtensionResourceOperation(resourceType, operationType, ...)
-const operationTypeArgIndexByDecorator: Record<GenericArmResourceDecorator, number> = {
-  $extensionResourceOperation: 2,
-  $legacyResourceOperation: 1,
-  $builtInResourceOperation: 2,
-  $legacyExtensionResourceOperation: 1,
-};
-
-const resourceOperationDecorators: { [verb in HttpVerb]: string[] } = {
-  put: ["$armResourceCreateOrUpdate", ...genericArmResourceDecorators],
-  get: ["$armResourceRead", "$armResourceList", ...genericArmResourceDecorators],
-  patch: ["$armResourceUpdate", ...genericArmResourceDecorators],
-  delete: ["$armResourceDelete", ...genericArmResourceDecorators],
-  post: ["$armResourceAction", "$armResourceCollectionAction", ...genericArmResourceDecorators],
-  head: [],
-};
-
-// ARM operationType conventions:
-// - PUT => createOrUpdate
-// - GET => read or list
-// - PATCH => update
-// - DELETE => delete
-// - POST => action
-// This map is used only for generic private decorators that encode operation type as an argument.
-const operationTypesByVerb: Partial<Record<HttpVerb, readonly string[]>> = {
-  put: ["createOrUpdate"],
-  get: ["read", "list"],
-  patch: ["update"],
-  delete: ["delete"],
-  post: ["action"],
-};
-
-function isGenericArmResourceDecorator(name: string): name is GenericArmResourceDecorator {
-  return genericArmResourceDecorators.includes(name as GenericArmResourceDecorator);
-}
-
-function getGenericDecoratorOperationType(decorator: DecoratorApplication): string | undefined {
-  const decoratorName = decorator.decorator.name;
-  if (!isGenericArmResourceDecorator(decoratorName)) {
-    return undefined;
-  }
-
-  const operationTypeArg = decorator.args[operationTypeArgIndexByDecorator[decoratorName]];
-  return getDecoratorArgumentStringValue(operationTypeArg);
-}
-
-function getDecoratorArgumentStringValue(arg: DecoratorArgument | undefined): string | undefined {
-  return typeof arg?.jsValue === "string" ? arg.jsValue : undefined;
-}
-
-function isStaticSegment(segment: string | undefined): boolean {
-  return segment === undefined || segment === "locations";
-}
-
-function isArmProviderOperation(program: Program, operation: Operation): boolean {
-  let isProviderAction = false;
-
-  for (const [index, [, property]] of [...operation.parameters.properties].entries()) {
-    const segment = getSegment(program, property);
-
-    if (segment === "providers") {
-      isProviderAction = true;
-    } else if (isProviderAction && !isStaticSegment(segment)) {
-      return false;
-    }
-
-    const isLastSegment = index === operation.parameters.properties.size - 1;
-    if (isLastSegment && !segment) {
-      return true;
-    }
-  }
-
-  return false;
-}
