--- packages/typespec-azure-resource-manager/test/rules/use-operation-decorator.test.ts
@@ -1,301 +1,0 @@
-import { Tester } from "#test/tester.js";
-import {
-  LinterRuleTester,
-  TesterInstance,
-  createLinterRuleTester,
-} from "@typespec/compiler/testing";
-import { beforeEach, describe, it } from "vitest";
-
-import { useOperationDecoratorRule } from "../../src/rules/use-operation-decorator.js";
-
-let runner: TesterInstance;
-let tester: LinterRuleTester;
-
-beforeEach(async () => {
-  runner = await Tester.createInstance();
-  tester = createLinterRuleTester(
-    runner,
-    useOperationDecoratorRule,
-    "@azure-tools/typespec-azure-resource-manager",
-  );
-});
-
-it("emits diagnostics if missing armResourceOperation decorators.", async () => {
-  await tester
-    .expect(
-      `
-        @armProviderNamespace
-        namespace Microsoft.Foo;
-        
-        model FooResource is TrackedResource<{}> {
-          ...ResourceNameParameter<FooResource>;
-        }
-
-        @armResourceOperations
-        interface FooResources extends ResourceCollectionOperations<FooResource> {
-          @put createOrUpdate( ...ResourceInstanceParameters<FooResource>, @bodyRoot resource: FooResource): ArmResponse<FooResource> | ArmCreatedResponse<FooResource> | ErrorResponse;
-          @get get(...ResourceInstanceParameters<FooResource>): ArmResponse<FooResource> | ErrorResponse;
-        #suppress "@typespec/http/deprecated-implicit-optionality" "For test"
-          @patch(#{implicitOptionality: true}) update(...ResourceInstanceParameters<FooResource>, @bodyRoot properties: Foundations.ResourceUpdateModel<FooResource, {}>): ArmResponse<FooResource> | ErrorResponse;
-          @delete delete(...ResourceInstanceParameters<FooResource>): | ArmDeletedResponse | ArmDeleteAcceptedResponse | ArmDeletedNoContentResponse | ErrorResponse;
-          @post action(...ResourceInstanceParameters<FooResource>) : ArmResponse<FooResource> | ErrorResponse;
-        }
-      `,
-    )
-    .toEmitDiagnostics([
-      {
-        code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-        message:
-          "Resource PUT operation must be decorated with @armResourceCreateOrUpdate or @extensionResourceOperation or @legacyResourceOperation or @builtInResourceOperation or @legacyExtensionResourceOperation.",
-      },
-      {
-        code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-        message:
-          "Resource GET operation must be decorated with @armResourceRead or @armResourceList or @extensionResourceOperation or @legacyResourceOperation or @builtInResourceOperation or @legacyExtensionResourceOperation.",
-      },
-      {
-        code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-        message:
-          "Resource PATCH operation must be decorated with @armResourceUpdate or @extensionResourceOperation or @legacyResourceOperation or @builtInResourceOperation or @legacyExtensionResourceOperation.",
-      },
-      {
-        code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-        message:
-          "Resource DELETE operation must be decorated with @armResourceDelete or @extensionResourceOperation or @legacyResourceOperation or @builtInResourceOperation or @legacyExtensionResourceOperation.",
-      },
-      {
-        code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-        message:
-          "Resource POST operation must be decorated with @armResourceAction or @armResourceCollectionAction or @extensionResourceOperation or @legacyResourceOperation or @builtInResourceOperation or @legacyExtensionResourceOperation.",
-      },
-    ]);
-});
-
-it("doesn't emit diagnostic for internal operations", async () => {
-  await tester
-    .expect(
-      `
-    @armProviderNamespace
-    namespace Microsoft.Foo;
-
-    model VmSize {
-      cpus: int32;
-    }
-
-    @armResourceOperations
-    interface ProviderOperations {
-      @get
-      getVmsSizes is ArmProviderActionSync<void, VmSize, SubscriptionActionScope>;
-    }
-    `,
-    )
-    .toBeValid();
-});
-
-it("doesn't emit diagnostic for operation templates", async () => {
-  await tester
-    .expect(
-      `
-      @armProviderNamespace
-      namespace Microsoft.ContosoProviderHub;
-
-      @armResourceOperations
-      interface Employees {
-        getData is ArmBodyProviderActionSync<string>;
-      }
-
-      op ArmBodyProviderActionSync<T>(t: T, ...Azure.ResourceManager.Foundations.DefaultBaseParameters<TenantActionScope>): void;
-    `,
-    )
-    .toBeValid();
-});
-
-describe("Private ARM resource decorators", () => {
-  function buildPrivateDecoratorTestSpec(operation: string = ""): string {
-    return `
-      @armProviderNamespace
-      namespace Microsoft.Foo;
-      using Azure.ResourceManager.Private;
-
-      model FooResource is TrackedResource<{}> {
-        ...ResourceNameParameter<FooResource>;
-      }
-
-      model BarResource is ExtensionResource<{}> {
-        ...ResourceNameParameter<BarResource>;
-      }
-
-      @armResourceOperations
-      interface FooResources {
-        ${operation}
-      }
-    `;
-  }
-
-  it("doesn't emit a diagnostic for operations with @extensionResourceOperation", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @extensionResourceOperation(FooResource, BarResource, "read")
-      @get
-      getExtension(...ResourceInstanceParameters<FooResource>): ArmResponse<BarResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toBeValid();
-  });
-
-  it("doesn't emit a diagnostic for operations with @legacyResourceOperation", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @legacyResourceOperation(FooResource, "read")
-      @get
-      getLegacy(...ResourceInstanceParameters<FooResource>): ArmResponse<FooResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toBeValid();
-  });
-
-  it("doesn't emit a diagnostic for operations with @builtInResourceOperation", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @builtInResourceOperation(FooResource, BarResource, "list")
-      @get
-      listBuiltIn(...ResourceInstanceParameters<FooResource>): ArmResponse<BarResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toBeValid();
-  });
-
-  it("doesn't emit a diagnostic for operations with @legacyExtensionResourceOperation", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @legacyExtensionResourceOperation(FooResource, "read")
-      @get
-      getLegacyExtension(...ResourceInstanceParameters<FooResource>): ArmResponse<FooResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toBeValid();
-  });
-
-  it("emits diagnostic when @extensionResourceOperation operationType doesn't match the HTTP verb", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @extensionResourceOperation(FooResource, BarResource, "createOrUpdate")
-      @post
-      doAction(...ResourceInstanceParameters<FooResource>): ArmResponse<BarResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toEmitDiagnostics({
-      code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-      message:
-        'Resource POST operation decorator @extensionResourceOperation must use operationType "action".',
-    });
-  });
-
-  it("emits diagnostic when @legacyResourceOperation operationType doesn't match the HTTP verb", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @legacyResourceOperation(FooResource, "createOrUpdate")
-      @post
-      doAction(...ResourceInstanceParameters<FooResource>): ArmResponse<FooResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toEmitDiagnostics({
-      code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-      message:
-        'Resource POST operation decorator @legacyResourceOperation must use operationType "action".',
-    });
-  });
-
-  it("emits diagnostic when @builtInResourceOperation operationType doesn't match the HTTP verb", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @builtInResourceOperation(FooResource, BarResource, "createOrUpdate")
-      @post
-      doAction(...ResourceInstanceParameters<FooResource>): ArmResponse<BarResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toEmitDiagnostics({
-      code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-      message:
-        'Resource POST operation decorator @builtInResourceOperation must use operationType "action".',
-    });
-  });
-
-  it("emits diagnostic when @legacyExtensionResourceOperation operationType doesn't match the HTTP verb", async () => {
-    const content = buildPrivateDecoratorTestSpec(`
-      @legacyExtensionResourceOperation(FooResource, "createOrUpdate")
-      @post
-      doAction(...ResourceInstanceParameters<FooResource>): ArmResponse<FooResource> | ErrorResponse;
-    `);
-    await tester.expect(content).toEmitDiagnostics({
-      code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-      message:
-        'Resource POST operation decorator @legacyExtensionResourceOperation must use operationType "action".',
-    });
-  });
-});
-
-describe("Provider operations", () => {
-  function providerOperationSetup(operation: string = "", content: string = ""): string {
-    return `
-      @armProviderNamespace
-      namespace Microsoft.Foo;
-  
-      model Employee is TrackedResource<{}> {
-        ...ResourceNameParameter<Employee>;
-      }
-
-      op ComputeProviderActionAsync<
-        Request extends {} | void = void,
-        Response extends {} | void = void,
-        Parameters extends {} = {},
-      > is ArmProviderActionAsync<Request, Response, TenantActionScope, Parameters>;
-       
-      ${content}
-
-      interface Operations extends Azure.ResourceManager.Operations {}
-
-      interface VirtualMachinesOperations {
-        ${operation}
-      }
-    `;
-  }
-
-  it("doesn't emit a diagnostic for provider operations", async () => {
-    const content = providerOperationSetup(
-      `
-        @autoRoute
-        @get
-        @action("virtualMachines")
-        listByLocation is ComputeProviderActionAsync<Response = ResourceListResult<Employee>>;
-        `,
-    );
-    await tester.expect(content).toBeValid();
-  });
-
-  it("doesn't emit a diagnostic for provider operations with location", async () => {
-    const content = providerOperationSetup(
-      `
-        @autoRoute
-        @get
-        @action("virtualMachinesGet")
-        get is ComputeProviderActionAsync<Response = ResourceListResult<Employee>, Parameters = LocationParameter>;
-        `,
-    );
-    await tester.expect(content).toBeValid();
-  });
-
-  it("emit diagnostic for provider operations if it has dynamic segments", async () => {
-    const content = providerOperationSetup(
-      `
-        @autoRoute
-        @get
-        @action("virtualMachines")
-        listByLocation is ComputeProviderActionAsync<
-          Response = ResourceListResult<Employee>,
-          Parameters = LocationParameter & MoveResponseParameter
-        >; 
-        `,
-      ` 
-        model MoveResponseParameter {
-          @path
-          @minLength(1)
-          @segment("moves")
-          move: string;
-        }
-        `,
-    );
-
-    await tester.expect(content).toEmitDiagnostics({
-      code: "@azure-tools/typespec-azure-resource-manager/use-operation-decorator",
-      message:
-        "Resource GET operation must be decorated with @armResourceRead or @armResourceList or @extensionResourceOperation or @legacyResourceOperation or @builtInResourceOperation or @legacyExtensionResourceOperation.",
-    });
-  });
-});
