+++ website/src/content/docs/docs/libraries/azure-resource-manager/reference/linter.md
@@ -39,8 +39,6 @@ Available ruleSets:
 | [`@azure-tools/typespec-azure-resource-manager/arm-resource-key-invalid-chars`](../rules/arm-resource-key-invalid-chars.md)                       | Arm resource key must contain only alphanumeric characters.                                                                                                                                                                                           |
 | [`@azure-tools/typespec-azure-resource-manager/arm-resource-name-pattern`](../rules/arm-resource-name-pattern.md)                                 | The resource name parameter should be defined with a 'pattern' restriction.                                                                                                                                                                           |
 | [`@azure-tools/typespec-azure-resource-manager/arm-resource-operation-response`](../rules/arm-resource-operation-response.md)                     | [RPC 008]: PUT, GET, PATCH & LIST must return the same resource schema.                                                                                                                                                                               |
-| [`@azure-tools/typespec-azure-resource-manager/use-api-version`](../rules/use-api-version.md)                                                     | Validate ARM Resource operations include the api-version parameter referencing Azure.ResourceManager.CommonTypes.ApiVersionParameter.                                                                                                                 |
-| [`@azure-tools/typespec-azure-resource-manager/use-operation-decorator`](../rules/use-operation-decorator.md)                                     | Validate ARM Resource operations use the correct decorator for the HTTP verb.                                                                                                                                                                         |
 | [`@azure-tools/typespec-azure-resource-manager/arm-resource-path-segment-invalid-chars`](../rules/arm-resource-path-segment-invalid-chars.md)     | Arm resource name must contain only alphanumeric characters.                                                                                                                                                                                          |
 | [`@azure-tools/typespec-azure-resource-manager/arm-resource-provisioning-state`](../rules/arm-resource-provisioning-state.md)                     | Check for properly configured provisioningState property.                                                                                                                                                                                             |
 | [`@azure-tools/typespec-azure-resource-manager/version-progression`](../rules/version-progression.md)                                             | Validate that ARM service versions all use unique dates and are declared in strictly increasing chronological order.                                                                                                                                  |
@@ -48,7 +46,7 @@ Available ruleSets:
 | [`@azure-tools/typespec-azure-resource-manager/arm-custom-resource-usage-discourage`](../rules/arm-custom-resource-usage-discourage.md)           | Verify the usage of @customAzureResource decorator.                                                                                                                                                                                                   |
 | [`@azure-tools/typespec-azure-resource-manager/arm-feature-file-usage-discourage`](../rules/arm-feature-file-usage-discourage.md)                 | Verify the usage of @featureFiles decorator.                                                                                                                                                                                                          |
 | [`@azure-tools/typespec-azure-resource-manager/beyond-nesting-levels`](../rules/beyond-nesting-levels.md)                                         | Tracked Resources must use 3 or fewer levels of nesting.                                                                                                                                                                                              |
-| [`@azure-tools/typespec-azure-resource-manager/use-interface`](../rules/use-interface.md)                                                         | Validate ARM Resource operations are defined inside an interface declaration.                                                                                                                                                                         |
+| [`@azure-tools/typespec-azure-resource-manager/arm-resource-operation`](../rules/arm-resource-operation.md)                                       | Validate ARM Resource operations.                                                                                                                                                                                                                     |
 | [`@azure-tools/typespec-azure-resource-manager/no-resource-delete-operation`](../rules/no-resource-delete-operation.md)                           | Check for resources that must have a delete operation.                                                                                                                                                                                                |
 | [`@azure-tools/typespec-azure-resource-manager/empty-updateable-properties`](../rules/empty-updateable-properties.md)                             | Should have updateable properties.                                                                                                                                                                                                                    |
 | [`@azure-tools/typespec-azure-resource-manager/arm-resource-interface-requires-decorator`](../rules/arm-resource-interface-requires-decorator.md) | Each resource interface must have an @armResourceOperations decorator.                                                                                                                                                                                |
